01-ai-volume-one/experiments/mcp/mcp-use/requirements.txt
mcp-use
langchain-openai
python-dotenv
openai
typing_extensions
langchain-anthropic
anthropic
pydantic
pydantic-core
typing_inspection
mcp>=1.5.0
langchain>=0.1.0
langchain-community>=0.0.10
websockets>=12.0
aiohttp>=3.9.0
jsonschema-pydantic>=0.1.0
annotated-types>=0.5.0
httpx>=0.23.0
idna